FE API: move order type extensions from project-base to the frontend-api package
