+++ packages/framework/src/Model/Order/OrderDataFactory.php
@@ -109,6 +109,7 @@ class OrderDataFactory
 
         $orderData->heurekaAgreement = $order->isHeurekaAgreement();
         $orderData->trackingNumber = $order->getTrackingNumber();
+        $orderData->gtmCoupon = $order->getGtmCoupon();
     }
 
     /**
