+++ packages/frontend-api/src/Model/Resolver/Order/OrderResolverMap.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Shopsys\FrontendApiBundle\Model\Resolver\Order;
 
+use Overblog\DataLoader\DataLoaderInterface;
 use Overblog\GraphQLBundle\Resolver\ResolverMap;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Model\Order\Order;
@@ -12,9 +13,12 @@ class OrderResolverMap extends ResolverMap
 {
     /**
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
+     * @param \Overblog\DataLoader\DataLoaderInterface $orderItemsBatchLoader
      */
-    public function __construct(protected readonly Domain $domain)
-    {
+    public function __construct(
+        protected readonly Domain $domain,
+        protected readonly DataLoaderInterface $orderItemsBatchLoader,
+    ) {
     }
 
     /**
@@ -24,20 +28,20 @@ class OrderResolverMap extends ResolverMap
     {
         return [
             'Order' => [
-                'country' => function (Order $order) {
-                    return $order->getCountry()->getCode();
-                },
                 'creationDate' => function (Order $order) {
                     return $order->getCreatedAt();
-                },
-                'deliveryCountry' => function (Order $order) {
-                    return $order->getDeliveryCountry() === null ? '' : $order->getDeliveryCountry()->getCode();
                 },
                 'isDeliveryAddressDifferentFromBilling' => function (Order $order) {
                     return !$order->isDeliveryAddressSameAsBillingAddress();
                 },
                 'status' => function (Order $order) {
                     return $order->getStatus()->getName($this->domain->getLocale());
+                },
+                'promoCode' => function (Order $order) {
+                    return $order->getGtmCoupon();
+                },
+                'items' => function (Order $order) {
+                    return $this->orderItemsBatchLoader->load($order);
                 },
             ],
         ];
