+++ packages/frontend-api/src/Resources/config/services.yaml
@@ -6,6 +6,7 @@ services:
             Overblog\DataLoader\DataLoaderInterface $imagesBatchLoader: '@images_batch_loader'
             Overblog\DataLoader\DataLoaderInterface $firstImageBatchLoader: '@first_image_batch_loader'
             Overblog\DataLoader\DataLoaderInterface $categoriesBatchLoader: '@categories_batch_loader'
+            Overblog\DataLoader\DataLoaderInterface $orderItemsBatchLoader: '@order_items_batch_loader'
         public: false
 
     Shopsys\FrontendApiBundle\:
