+++ project-base/app/config/graphql/types/ModelType/Order/Order.types.yaml
@@ -2,23 +2,3 @@ Order:
     type: object
     inherits:
         - 'OrderDecorator'
-    config:
-        fields:
-            firstName:
-                type: "String"
-                description: "The customer's first name"
-            lastName:
-                type: "String"
-                description: "The customer's last name"
-            promoCode:
-                type: "String"
-                description: "Promo code (coupon) used in the order"
-            country:
-                type: "Country!"
-                description: "Billing address country"
-            deliveryCountry:
-                type: "Country"
-                description: "Country for delivery"
-            productItems:
-                type: "[OrderItem!]!"
-                description: "All product items in the order"
