+++ project-base/app/config/packages/overblog_dataloader.yaml
@@ -28,7 +28,7 @@ overblog_dataloader:
             batch_load_fn: "@Shopsys\\LuigisBoxBundle\\Model\\Batch\\LuigisBoxBatchLoader:loadByBatchData"
         orderItemsBatchLoader:
             alias: "order_items_batch_loader"
-            batch_load_fn: "@App\\FrontendApi\\Model\\Order\\OrderItemsBatchLoader:loadAllByOrders"
+            batch_load_fn: "@Shopsys\\FrontendApiBundle\\Model\\Order\\OrderItemsBatchLoader:loadAllByOrders"
         productsByEntitiesBatchLoader:
             alias: "products_by_entities_batch_loader"
             batch_load_fn: "@App\\FrontendApi\\Model\\Product\\ProductsBatchLoader:loadByEntities"
