+++ project-base/app/src/Model/Order/OrderDataFactory.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Model\Order;
 
-use Shopsys\FrameworkBundle\Model\Order\Order as BaseOrder;
 use Shopsys\FrameworkBundle\Model\Order\OrderData as BaseOrderData;
 use Shopsys\FrameworkBundle\Model\Order\OrderDataFactory as BaseOrderDataFactory;
 
@@ -14,6 +13,7 @@ use Shopsys\FrameworkBundle\Model\Order\OrderDataFactory as BaseOrderDataFactory
  * @method \App\Model\Order\OrderData create()
  * @method \App\Model\Order\OrderData createFromOrder(\App\Model\Order\Order $order)
  * @method \App\Model\Order\OrderData fillZeroPrices(\App\Model\Order\OrderData $orderData)
+ * @method fillFromOrder(\App\Model\Order\OrderData $orderData, \App\Model\Order\Order $order)
  */
 class OrderDataFactory extends BaseOrderDataFactory
 {
@@ -23,16 +23,5 @@ class OrderDataFactory extends BaseOrderDataFactory
     protected function createInstance(): BaseOrderData
     {
         return new OrderData();
-    }
-
-    /**
-     * @param \App\Model\Order\OrderData $orderData
-     * @param \App\Model\Order\Order $order
-     */
-    protected function fillFromOrder(BaseOrderData $orderData, BaseOrder $order): void
-    {
-        parent::fillFromOrder($orderData, $order);
-
-        $orderData->gtmCoupon = $order->getGtmCoupon();
     }
 }
